day 11 part 2

+++ 2022/Day11/Models/Item.cs
@@ -2,5 +2,5 @@
 
 public record Item
 {
-	public int WorryLevel { get; set; }
+	public long WorryLevel { get; set; }
 }
